requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
courlan==1.0.0
cssselect==1.2.0
dateparser==1.2.0
environ==1.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==3.0.2
goose3==3.1.19
htmldate==1.7.0
idna==3.6
itsdangerous==2.1.2
jieba==0.42.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jusText==3.0.0
langcodes==3.3.0
langdetect==1.0.9
lxml==5.1.0
markdownify==0.11.6
MarkupSafe==2.1.5
newspaper3k==0.2.8
nltk==3.8.1
pillow==10.2.0
pyahocorasick==2.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
setuptools==69.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tld==0.13
tldextract==5.1.1
tqdm==4.66.2
trafilatura==1.7.0
tzlocal==5.2
urllib3==2.2.0
Werkzeug==3.0.1
